requirements.txt
pinecone-client[grpc]
datasets
sentence-transformers